requirements.txt
asgiref==3.5.0
certifi==2022.6.15
charset-normalizer==2.1.0
distlib==0.3.6
dj-database-url==1.0.0
Django==4.0.2
django-ckeditor==6.2.0
django-cors-headers==3.13.0
django-filter==21.1
django-js-asset==2.0.0
djangorestframework==3.14.0
filelock==3.8.0
gcloud==0.18.3
googleapis-common-protos==1.56.0
gunicorn==20.1.0
httplib2==0.20.4
idna==3.3
jws==0.1.3
oauth2client==4.1.3
Pillow==9.0.1
platformdirs==2.5.2
protobuf==3.20.0rc2
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.15.0
pyparsing==3.0.7
Pyrebase4==4.5.0
python-jwt==2.0.1
pytz==2022.2.1
requests==2.28.1
requests-toolbelt==0.9.1
rsa==4.8
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.11
virtualenv==20.16.5
whitenoise==6.2.0
